pyproject.toml
[tool.poetry]
name = "storyloop"
version = "0.4.0"
description = "AI-powered collaborative visual storytelling app"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.3"
Flask-Login = "^0.6.3"
Flask-WTF = "^1.2.1"
Flask-SQLAlchemy = "^3.1.1"
Flask-Migrate = "^4.0.5"
Flask-SocketIO = "^5.3.6"
email-validator = "^2.1.1"
celery = "^5.4.0"
redis = "^5.0.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9"
# Google Cloud
google-cloud-aiplatform = {version = ">=1.55.0", extras = ["preview"]}
google-cloud-storage = "^2.16.0"
ffmpeg-python = "^0.2.0"
eventlet = "^0.35.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.4.0"
flake8 = "^7.0.0"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"